feat: :sparkles: Add smooth scrolling to anchors
